@@ -1,73 +1,43 @@
 {
-	"name": "react-template",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^11.2.7",
-		"@testing-library/user-event": "^12.8.3",
-		"@types/jest": "^27.0.2",
-		"@types/node": "^12.20.4",
-		"@types/react": "^17.0.31",
-		"@types/react-dom": "^17.0.10",
-		"@types/react-router-dom": "^5.3.2",
-		"axios": "^0.23.0",
-		"bulma": "^0.9.3",
-		"history": "^5.0.1",
-		"node-sass": "^6.0.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-router": "^6.0.0-beta.8",
-		"react-router-dom": "^6.0.0-beta.8",
-		"react-scripts": "4.0.3",
-		"typescript": "^4.4.4",
-		"web-vitals": "^1.1.2"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"lint": "eslint --fix \"src/**/*.{ts,tsx}\"",
-		"prettier": "prettier --check \"src/**/*.{ts,tsx}\"",
-		"storybook": "start-storybook -p 6006 -s public",
-		"prebuild:storybook": "npm run test:generate-output",
-		"build-storybook": "build-storybook -s public",
-		"test:generate-output": "jest --json --outputFile=src/.test-results.json --watchAll",
-		"prepare": "husky install"
-	},
-	"eslintConfig": {
-		"extends": [
-			"react-app",
-			"react-app/jest"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"@storybook/addon-a11y": "^6.3.12",
-		"@storybook/addon-actions": "^6.3.12",
-		"@storybook/addon-essentials": "^6.3.12",
-		"@storybook/addon-jest": "^6.3.12",
-		"@storybook/addon-knobs": "^6.3.1",
-		"@storybook/addon-links": "^6.3.12",
-		"@storybook/node-logger": "^6.3.12",
-		"@storybook/preset-create-react-app": "^3.2.0",
-		"@storybook/react": "^6.3.12",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"husky": "^7.0.0",
-		"prettier": "^2.4.1",
-		"pretty-quick": "^3.1.1"
-	}
+  "name": "react-template",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.0",
+    "@types/node": "^16.11.33",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.6.4",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
