@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import App from './App';
 
 test('renders learn react link', () => {
-	render(<App />);
-	const linkElement = screen.getByText(/React template!/i);
-	expect(linkElement).toBeInTheDocument();
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
 });
